% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{AirportDelayVisualizer}
\alias{AirportDelayVisualizer}
\title{Airport Delay Visualizer}
\description{
This R6 class provides a visual representation of the mean delay of flights for different airports,
             based on longitude and latitude information. It uses data from the `nycflights13` package.
}
\examples{

## ------------------------------------------------
## Method `AirportDelayVisualizer$new`
## ------------------------------------------------

visualizer <- AirportDelayVisualizer$new()
visualizer$plot()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{airport_data}}{Data regarding airports with mean delay information and geographical location.}

\item{\code{delay_data}}{Data regarding mean arrival delays by airport destination.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AirportDelayVisualizer-new}{\code{AirportDelayVisualizer$new()}}
\item \href{#method-AirportDelayVisualizer-prepare_data}{\code{AirportDelayVisualizer$prepare_data()}}
\item \href{#method-AirportDelayVisualizer-plot}{\code{AirportDelayVisualizer$plot()}}
\item \href{#method-AirportDelayVisualizer-clone}{\code{AirportDelayVisualizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AirportDelayVisualizer-new"></a>}}
\if{latex}{\out{\hypertarget{method-AirportDelayVisualizer-new}{}}}
\subsection{Method \code{new()}}{
Initialize the AirportDelayVisualizer class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AirportDelayVisualizer$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object of class AirportDelayVisualizer.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{visualizer <- AirportDelayVisualizer$new()
visualizer$plot()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AirportDelayVisualizer-prepare_data"></a>}}
\if{latex}{\out{\hypertarget{method-AirportDelayVisualizer-prepare_data}{}}}
\subsection{Method \code{prepare_data()}}{
Prepare Data
Calculates the mean delays by airport destination and merges the data with
airport location information (longitude and latitude).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AirportDelayVisualizer$prepare_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AirportDelayVisualizer-plot"></a>}}
\if{latex}{\out{\hypertarget{method-AirportDelayVisualizer-plot}{}}}
\subsection{Method \code{plot()}}{
Plot Mean Airport Delays
Creates a scatter plot using `ggplot2` to visualize mean flight delays at
different airport locations. The size and color of each point represent the
mean delay in minutes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AirportDelayVisualizer$plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot object displaying mean delays at airports by location.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AirportDelayVisualizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AirportDelayVisualizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AirportDelayVisualizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
