% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack Bruteforce Approach}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data.frame consisting of weights and values}

\item{W}{size of the knapsack}
}
\value{
optimal knapsack solution
}
\description{
This function implements a solution to the knapsack problem through brute force.
This means it calculates every single possible solution (2^n) and chooses the best.
This is achieved through running a loop 2^n times, 
converting the loop variable to a binary representation, using that to slice the 
elements and checking whether the solution is valid and whether it  is better than all prior solutions.
}
\examples{
knapsack_objects <- data.frame(
  w = sample(1:4000, size = 16, replace = TRUE),
  v = runif(n = 16, min = 0, max = 10000)
  )
result <- brute_force_knapsack(x = knapsack_objects, W = 3500)
}
