% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{First integer}

\item{y}{Second Integer}
}
\value{
numeric value
}
\description{
The euclidean algorithm determines the largest common divisor of two integers.
It was derive by greek mathematician Euclid.
}
\examples{
x <- 121
y <- 11
euclidean(x,y)
}
