% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Knappsack Greedy Implementation}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame consisting of weights and values}

\item{W}{size of the knapsack}

\item{name}{description}
}
\value{
decent knapsack solution
}
\description{
This method implements a greedy approach to the knapsack problem.
It works by filling the knapsack S1 with the "best bang for the buck" (value/weight) ratio
items until at full capacity. It also checks whether the first not fitting item would 
constitute a better solution. It might not have a ratio as good as other items,
but might leave less empty space in the knapsack and ultimately would be a better solution than
S1. 
The algorithm does not guarantee an optimal solution, but a solution that is at least 50%
that of the other algorithms solutions while having a complexity of only O(nlogn).
}
\examples{
knapsack_objects <- data.frame(
  w = sample(1:4000, size = 16, replace = TRUE),
  v = runif(n = 16, min = 0, max = 10000)
  )
result <- greedy_knapsack(x = knapsack_objects, W = 3500)
}
