% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{KoladaHandler}
\alias{KoladaHandler}
\title{Kolada Handler}
\description{
An R6 class to handle data retrieval from Kolada API
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KoladaHandler-new}{\code{KoladaHandler$new()}}
\item \href{#method-KoladaHandler-parse_response}{\code{KoladaHandler$parse_response()}}
\item \href{#method-KoladaHandler-parse_kpi}{\code{KoladaHandler$parse_kpi()}}
\item \href{#method-KoladaHandler-get_data}{\code{KoladaHandler$get_data()}}
\item \href{#method-KoladaHandler-clone}{\code{KoladaHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoladaHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-KoladaHandler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoladaHandler$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoladaHandler-parse_response"></a>}}
\if{latex}{\out{\hypertarget{method-KoladaHandler-parse_response}{}}}
\subsection{Method \code{parse_response()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoladaHandler$parse_response(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoladaHandler-parse_kpi"></a>}}
\if{latex}{\out{\hypertarget{method-KoladaHandler-parse_kpi}{}}}
\subsection{Method \code{parse_kpi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoladaHandler$parse_kpi(kpi_string)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoladaHandler-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-KoladaHandler-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoladaHandler$get_data(kpi_ids, municipality_ids, year)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoladaHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KoladaHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoladaHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
