\name{lab3victorchris-package}
\alias{lab3victorchris-package}
\alias{lab3victorchris}
\docType{package}
\title{
\packageTitle{lab3victorchris}
}
\description{
\packageDescription{lab3victorchris}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{lab3victorchris}

Maintainer: \packageMaintainer{lab3victorchris}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = lab3victorchris)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{lab3victorchris}}
%\section{Documentation Index}{\packageIndices{lab3victorchris}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
