% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression R6 Class}
\description{
Linear Regression R6 Class

Linear Regression R6 Class
}
\details{
This R6 class performs linear regression using ordinary least squares.
}
\examples{
data(iris)
linreg_model <- linreg$new(Petal.Length ~ Species, iris)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formula}}{A formula object describing the model.}

\item{\code{data}}{A data frame containing the variables in the model.}

\item{\code{data_name}}{contains name of data set}

\item{\code{X}}{The model matrix generated from the formula and data.}

\item{\code{y}}{The response variable (dependent variable).}

\item{\code{beta_hat}}{The estimated regression coefficients.}

\item{\code{y_hat}}{The fitted values.}

\item{\code{e_hat}}{The residuals.}

\item{\code{df}}{Degrees of freedom of the model.}

\item{\code{sigma_squared}}{Residual variance of the model.}

\item{\code{var_beta_hat}}{Variance of the regression coefficients.}

\item{\code{t_beta}}{t-values for each coefficient.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-linreg-new}{\code{linreg$new()}}
\item \href{#method-linreg-print}{\code{linreg$print()}}
\item \href{#method-linreg-pred}{\code{linreg$pred()}}
\item \href{#method-linreg-resid}{\code{linreg$resid()}}
\item \href{#method-linreg-coef}{\code{linreg$coef()}}
\item \href{#method-linreg-plot}{\code{linreg$plot()}}
\item \href{#method-linreg-summary}{\code{linreg$summary()}}
\item \href{#method-linreg-clone}{\code{linreg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linreg-new"></a>}}
\if{latex}{\out{\hypertarget{method-linreg-new}{}}}
\subsection{Method \code{new()}}{
Constructor method for linear regression model using QR decomposition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{linreg$new(formula, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{A formula object that describes the model}

\item{\code{data}}{A data.frame containing the data on which the linear regression is performed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6 object representing the linear regression model
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linreg-print"></a>}}
\if{latex}{\out{\hypertarget{method-linreg-print}{}}}
\subsection{Method \code{print()}}{
Prints the coefficients in an orderly manner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{linreg$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linreg-pred"></a>}}
\if{latex}{\out{\hypertarget{method-linreg-pred}{}}}
\subsection{Method \code{pred()}}{
Returns the values that the model predicted
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{linreg$pred()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linreg-resid"></a>}}
\if{latex}{\out{\hypertarget{method-linreg-resid}{}}}
\subsection{Method \code{resid()}}{
Returns the model's residuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{linreg$resid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linreg-coef"></a>}}
\if{latex}{\out{\hypertarget{method-linreg-coef}{}}}
\subsection{Method \code{coef()}}{
Returns the model's coefficients as a named vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{linreg$coef()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linreg-plot"></a>}}
\if{latex}{\out{\hypertarget{method-linreg-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the residuals vs the fitted values using ggplot2
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{linreg$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linreg-summary"></a>}}
\if{latex}{\out{\hypertarget{method-linreg-summary}{}}}
\subsection{Method \code{summary()}}{
Summary of the linear regression model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{linreg$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linreg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-linreg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{linreg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
