% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{map_handler}
\alias{map_handler}
\title{Map Handler}
\description{
Handler class for map plotting
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{shapefile_data}}{A data frame containing shapefile data}

\item{\code{shapefile_path}}{The file path to the shapefile}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-map_handler-load_shapefile}{\code{map_handler$load_shapefile()}}
\item \href{#method-map_handler-new}{\code{map_handler$new()}}
\item \href{#method-map_handler-merge_data}{\code{map_handler$merge_data()}}
\item \href{#method-map_handler-plot_data}{\code{map_handler$plot_data()}}
\item \href{#method-map_handler-clone}{\code{map_handler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-load_shapefile"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-load_shapefile}{}}}
\subsection{Method \code{load_shapefile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map_handler$load_shapefile(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-new"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map_handler$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-merge_data"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-merge_data}{}}}
\subsection{Method \code{merge_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map_handler$merge_data(kolada_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map_handler$plot_data(merged_data, title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map_handler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
