% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MapHandler}
\alias{MapHandler}
\title{Map Handler}
\description{
Handler class for map plotting
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{shapefile_data}}{A data frame containing shapefile data}

\item{\code{shapefile_path}}{The file path to the shapefile}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-map_handler-load_shapefile}{\code{MapHandler$load_shapefile()}}
\item \href{#method-map_handler-new}{\code{MapHandler$new()}}
\item \href{#method-map_handler-merge_data}{\code{MapHandler$merge_data()}}
\item \href{#method-map_handler-plot_data}{\code{MapHandler$plot_data()}}
\item \href{#method-map_handler-highlight_municipality}{\code{MapHandler$highlight_municipality()}}
\item \href{#method-map_handler-empty_map}{\code{MapHandler$empty_map()}}
\item \href{#method-map_handler-clone}{\code{MapHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-load_shapefile"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-load_shapefile}{}}}
\subsection{Method \code{load_shapefile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MapHandler$load_shapefile(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-new"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MapHandler$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-merge_data"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-merge_data}{}}}
\subsection{Method \code{merge_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MapHandler$merge_data(kolada_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MapHandler$plot_data(merged_data = NULL, title = "Map")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-highlight_municipality"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-highlight_municipality}{}}}
\subsection{Method \code{highlight_municipality()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MapHandler$highlight_municipality(municipality_name, municipality_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-empty_map"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-empty_map}{}}}
\subsection{Method \code{empty_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MapHandler$empty_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_handler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-map_handler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MapHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
