% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikipedia Graph Example Dataset}
\format{
A data frame with 15 rows and 3 variables:
\describe{
  \item{v1}{Start node of an edge (integer)}
  \item{v2}{End node of an edge (integer)}
  \item{w}{Weight of the edge between the nodes (numeric)}
}
}
\source{
Wikipedia page on Dijkstra's algorithm: \url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
\usage{
wiki_graph
}
\description{
This dataset is a small example graph based on the Wikipedia page for 
Dijkstra's algorithm.
}
\examples{
data(wiki_graph)
print(wiki_graph)
}
\keyword{datasets}
